Add hypertrace workflow
This introduces the "Hypertrace" wrapper around Isofit, which takes a
known surface reflectance and atmospheric conditions, performs a forward
simulation of TOA radiance, and then inverts the simulated radiance to
estimate surface reflectance. The wrapper tries to defer to Isofit as
much as possible -- mostly, it generates the necessary Isofit config
files and does a few additional checks, and does very little in the way
of pre- or post-processing.

I've also added a GitHub Actions integration test that runs the workflow
over a range of conditions.